removed (obsolete) dataFR (~M_0_reg)
in a very early version, I could choose to display regression values
(with many regressors) instead of the cell fluorescence on the left
half, and cluster and explore subsequently. This has been too removed
from all other analyses and have been obsolete for some time. Should
write separate script/pop-up window if needed again and it shouldn't be
kept as modular as possible.