v1.2, raw first version with time-lists, basics working
Major change of data-structure: indexing not only by cell but by time
also. Now time-lists are preloaded with CONST.m and grouped by stimulus.
So you can interactively select a subgroup of stimuli for analysis. The
M_0_fluo copy is eliminated, M_0 and M copies are only local to a
function call, so memory load of the larger (newer) datasets should be
improved. e.g. now also downsampling before passing to display. Still
need to fix stim-lock, reformat the (everchanging?!) preload3
(especially the old 'datamenu'), and there is a bug in the dot
stimulus-averaging (but dots are not eliciting activity in this fish
anyways), and 'isRawtime' doesn't work yet.