Incorporated MotorSeed Regressors into GUI
added isMotorseed flag to GUI, option checked by GetMotorRegressor to
ouput either the motorseed regressors or the old fictive trace convolved
regressor. Also GetTimeIndexedData updates the 'behavior' var that is
drawn on GUI and also passed around. The motorseeds are generated by
fig3_motormap (which is a scripted procedure inspired by manual
discovery), saved into VAR (11,1) and then the full length avr
functional trace of these seed cells are saved into GUIdata as
Behavior_full_motorseed and BehaviorAvr_motorseed. This data is not
saved in ('nested') 'dataset', since it is downstream of VAR.